Test improvements to reduce run time.
