+++ sdk/eventhub/Microsoft.Azure.EventHubs/tests/Processor/ProcessorNegativeCases.cs
@@ -10,7 +10,6 @@ namespace Microsoft.Azure.EventHubs.Tests.Processor
 
     public class NegativeCases : ProcessorTestBase
     {
-
         [Fact]
         [LiveTest]
         [DisplayTestMethodName]
